infra/requirements.txt
alembic==1.7.3
attrs==21.2.0
black==21.9b0
click==8.0.1
flasgger==0.9.5
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.0
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.8.1
importlib-resources==5.2.2
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==4.0.1
Mako==1.1.5
MarkupSafe==2.0.1
mistune==0.8.4
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
psycopg2-binary==2.9.1
pydantic==1.8.2
PyJWT==2.2.0
pyrsistent==0.18.0
python-dotenv==0.19.0
PyYAML==5.4.1
regex==2021.9.30
six==1.16.0
SQLAlchemy==1.4.25
tomli==1.2.1
typed-ast==1.4.3
typing-extensions==3.10.0.2
ujson==4.2.0
Werkzeug==2.0.1
zipp==3.6.0
apispec==5.1.1